+++ packages/core-js/stable/promise/with-resolvers.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/promise/with-resolvers');
+
+module.exports = parent;
