+++ packages/core-js/stage/4.js
@@ -8,6 +8,7 @@ require('../proposals/change-array-by-copy-stage-4');
 require('../proposals/global-this');
 require('../proposals/promise-all-settled');
 require('../proposals/promise-any');
+require('../proposals/promise-with-resolvers');
 require('../proposals/relative-indexing-method');
 require('../proposals/string-match-all');
 require('../proposals/string-replace-all-stage-4');
