+++ tests/unit-global/es.promise.with-resolvers.js
